Add in more of the necessary logic on the publish side.
That is, make sure to deal with inter-process publishing properly.
This requires us to introduce two copies of
do_intra_process_and_return_shared(), one of which deals with the
ROS Message type and the other that deals with the PublishedType.
This is kind of wasteful, and we may get rid of this later on,
but this works for now.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>