Fixed bugs, fixed linters
Fixed image tools issues

Signed-off-by: Gonzalo de Pedro <gonzalo@depedro.com.ar>

Fix linting. (#12)

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

Fix the test for any_subscription_callback. (#13)

In particular, change it so that the tests for TypeAdaptation
with any_subscription_callback always pass the custom type
to dispatch_intra_process().  That means that when using
TypeAdaptation, it is not possible to call dispatch_intra_process()
with the ROS message type.  But since this is a low-level
interface, this should be fine; the public-facing APIs handle
this case for us.

This also requires us to remove the test for inter-process
publishing with type adaptation.  That's because we didn't change
the signature of AnySubscriptionCallback::dispath() to take
in the custom type, so it always expects the ROS message type.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

Reenable the test_intra_process* tests. (#14)

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

Remove debugging print statements. (#11)

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>
Signed-off-by: Gonzalo de Pedro <gonzalo@depedro.com.ar>